+++ src/main/java/com/sh/sh_xiao_cheng_xu/service/ReportService.java
@@ -6,6 +6,6 @@ import com.sh.sh_xiao_cheng_xu.entity.ReportInfo;
 import java.util.List;
 
 public interface ReportService extends IService<ReportInfo> {
-    String publishReport(Long userId, Integer societyId, Long eventId, String desc, String imgUrls);
+    String publishReport(Long userId, Long societyId, Long eventId, String desc, String imgUrls);
     List<ReportInfo> getHistoryBySociety(Long societyId);
 }
